Add status to items that fail.
Great plugin! But I thought the addition of a status might be a useful addition if you're  using it to create a new short YOURL if the one being updated doesn't exist. ```"fail"``` is the standard YOURLS status for things that, well, fail, so I thought better to test for one word than a variety of status codes. If I get a "fail" I just make a new URL then. 

This may be a crap suggestion, feel free to reject, but it's working for me, and thanks for a great plugin!